♻️ refactor: 식당별 영업시간 조회 restaurant 도메인으로 이동
