+++ src/main/java/com/sparta/goodbite/domain/operatinghour/controller/OperatingHourController.java
@@ -1,18 +1,14 @@
 package com.sparta.goodbite.domain.operatinghour.controller;
 
-import com.sparta.goodbite.common.response.DataResponseDto;
 import com.sparta.goodbite.common.response.MessageResponseDto;
 import com.sparta.goodbite.common.response.ResponseUtil;
 import com.sparta.goodbite.domain.operatinghour.dto.CreateOperatingHourRequestDto;
-import com.sparta.goodbite.domain.operatinghour.dto.OperatingHourResponseDto;
 import com.sparta.goodbite.domain.operatinghour.dto.UpdateOperatingHourRequestDto;
 import com.sparta.goodbite.domain.operatinghour.service.OperatingHourService;
 import jakarta.validation.Valid;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -50,13 +46,5 @@ public class OperatingHourController {
 
         operatingHourService.deleteOperatingHour(operatingHourId);
         return ResponseUtil.deleteOk();
-    }
-
-    @GetMapping("/{restaurantId}")
-    public ResponseEntity<DataResponseDto<List<OperatingHourResponseDto>>> getAllOperatingHoursByRestaurant(
-        @PathVariable Long restaurantId) {
-
-        return ResponseUtil.findOk(
-            operatingHourService.getAllOperatingHoursByRestaurant(restaurantId));
     }
 }
